Move key types to central location
