+++ libs/common/src/auth/abstractions/password-reset-enrollment.service.abstraction.ts
@@ -1,4 +1,4 @@
-import { UserKey } from "../../platform/models/domain/symmetric-crypto-key";
+import { UserKey } from "../../types/key";
 
 export abstract class PasswordResetEnrollmentServiceAbstraction {
   /*
