+++ libs/common/src/auth/models/view/webauthn-login/webauthn-login-credential-assertion.view.ts
@@ -1,4 +1,4 @@
-import { PrfKey } from "../../../../platform/models/domain/symmetric-crypto-key";
+import { PrfKey } from "../../../../types/key";
 import { WebAuthnLoginAssertionResponseRequest } from "../../../services/webauthn-login/request/webauthn-login-assertion-response.request";
 
 export class WebAuthnLoginCredentialAssertionView {
