+++ libs/common/src/auth/services/webauthn-login/webauthn-login-prf-crypto.service.ts
@@ -1,5 +1,6 @@
 import { CryptoFunctionService } from "../../../platform/abstractions/crypto-function.service";
-import { PrfKey, SymmetricCryptoKey } from "../../../platform/models/domain/symmetric-crypto-key";
+import { SymmetricCryptoKey } from "../../../platform/models/domain/symmetric-crypto-key";
+import { PrfKey } from "../../../types/key";
 import { WebAuthnLoginPrfCryptoServiceAbstraction } from "../../abstractions/webauthn/webauthn-login-prf-crypto.service.abstraction";
 
 const LoginWithPrfSalt = "passwordless-login";
