+++ libs/common/src/platform/models/domain/account-keys.spec.ts
@@ -1,9 +1,10 @@
 import { makeStaticByteArray } from "../../../../spec";
 import { CsprngArray } from "../../../types/csprng";
+import { DeviceKey } from "../../../types/key";
 import { Utils } from "../../misc/utils";
 
 import { AccountKeys, EncryptionPair } from "./account";
-import { DeviceKey, SymmetricCryptoKey } from "./symmetric-crypto-key";
+import { SymmetricCryptoKey } from "./symmetric-crypto-key";
 
 describe("AccountKeys", () => {
   describe("toJSON", () => {
