+++ libs/common/src/platform/models/domain/symmetric-crypto-key.ts
@@ -1,4 +1,4 @@
-import { Jsonify, Opaque } from "type-fest";
+import { Jsonify } from "type-fest";
 
 import { Utils } from "../../../platform/misc/utils";
 import { EncryptionType } from "../../enums";
@@ -75,13 +75,3 @@ export class SymmetricCryptoKey {
     return SymmetricCryptoKey.fromString(obj?.keyB64);
   }
 }
-
-// Setup all separate key types as opaque types
-export type DeviceKey = Opaque<SymmetricCryptoKey, "DeviceKey">;
-export type PrfKey = Opaque<SymmetricCryptoKey, "PrfKey">;
-export type UserKey = Opaque<SymmetricCryptoKey, "UserKey">;
-export type MasterKey = Opaque<SymmetricCryptoKey, "MasterKey">;
-export type PinKey = Opaque<SymmetricCryptoKey, "PinKey">;
-export type OrgKey = Opaque<SymmetricCryptoKey, "OrgKey">;
-export type ProviderKey = Opaque<SymmetricCryptoKey, "ProviderKey">;
-export type CipherKey = Opaque<SymmetricCryptoKey, "CipherKey">;
